Reintroduce webworker with transferable objects.
